Initial branch insert for Visual Studio 2008 project files (and adjusted code re-order for pre-compiled headers)

+++ aphavoc/source/eechini.c
@@ -111,7 +111,7 @@ static void get_float ( const struct config_option *option, char *value )
 
 static void set_string ( const struct config_option *option, const char *value )
 {
-	if ( value && *value && strlen ( value ) < option->flag_length )
+	if ( value && *value && strlen ( value ) < (unsigned int) option->flag_length )
 	{
 		strcpy ( option->str_value, value );
 	}
@@ -281,7 +281,7 @@ static void get_mfd_pos ( const struct config_option *option, char *value )
 
 static void set_sound_device ( const struct config_option *option, const char *value )
 {
-	if ( value && *value && strlen ( value ) < option->flag_length )
+	if ( value && *value && strlen ( value ) < (unsigned int) option->flag_length )
 	{
 		int
 			i;
@@ -341,7 +341,7 @@ static void get_autosave ( const struct config_option *option, char *value )
 
 static void set_wut ( const struct config_option *option, const char *value )
 {
-	if ( value && *value && strlen ( value ) < option->flag_length )
+	if ( value && *value && strlen ( value ) < (unsigned int) option->flag_length )
 	{
 		strcpy ( option->str_value, value );
 
@@ -353,7 +353,7 @@ static void set_wut ( const struct config_option *option, const char *value )
 
 static void set_ip_address ( const struct config_option *option, const char *value )
 {
-	if ( value && *value && strlen ( value ) < option->flag_length )
+	if ( value && *value && strlen ( value ) < (unsigned int) option->flag_length )
 	{
 		strcpy ( option->str_value, value );
 
@@ -454,7 +454,7 @@ static void set_cfloat ( const struct config_option *option, const char *value )
 
 static void set_cstring ( const struct config_option *option, const char *value )
 {
-	if ( value && *value && strlen ( value ) < option->flag_length )
+	if ( value && *value && strlen ( value ) < (unsigned int) option->flag_length )
 	{
 		strcpy ( option->str_value, value );
 	}
